user index

+++ db/migrate/20160520174145_create_users.rb
@@ -5,9 +5,9 @@ class CreateUsers < ActiveRecord::Migration
       t.string :email
       t.string :roll
       t.string :phone
-      t.integer :cgpa
-      t.integer :x_th
-      t.integer :xii_th
+      t.float :cgpa
+      t.float :x_th
+      t.float :xii_th
       t.belongs_to :branch, index:true
       t.timestamps null: false
     end
